+++ VCore/Views and ViewControllers/Base Button/SwiftUI/SwiftUIBaseButtonState.swift
@@ -1,0 +1,36 @@
+//
+//  SwiftUIBaseButtonState.swift
+//  VCore
+//
+//  Created by Vakhtang Kontridze on 12/26/21.
+//
+
+import Foundation
+
+// MARK: - Base Button State
+/// Enum that describes state, such as `enabled` or `disabled`.
+public enum SwiftUIBaseButtonState: Int, CaseIterable {
+    // MARK: Cases
+    /// Enabled.
+    case enabled
+    
+    /// Disabled.
+    case disabled
+    
+    // MARK: Properties
+    /// Indicates if state is enabled.
+    public var isEnabled: Bool {
+        switch self {
+        case .enabled: return true
+        case .disabled: return false
+        }
+    }
+    
+    // MARK: Initializers
+    init(isEnabled: Bool) {
+        switch isEnabled {
+        case false: self = .disabled
+        case true: self = .enabled
+        }
+    }
+}