+++ CHANGELOG.md
@@ -6,8 +6,9 @@
 - Framework now supports `iOS 13`
 - `NetworkClient` now supports async/await
 - `MultiPartFormDataBuilder` has been added
-- `InfiniteScrollingCollectionView` has been added
+- `UIKitBaseButton` and `SwiftUIBaseButton` have been added
 - `KeyboardResponsiveViewController` has been added
+- `InfiniteScrollingCollectionView` has been added
 - `EdgeInsets` objects now support insetting and addition/subtraction
 - Several cases have been added to `NetworkError`
 - `UITableView` and `UICollectionView` dequeue and configure methods would no crash if a cell identifier is not registered